+++ addon/templates/components/em-input.hbs
@@ -1,5 +1,5 @@
 {{#em-form-group model=model property=property label=label form=form
-  elementClass=elementClass controlWrapper=controlWrapper labelClass=labelClass
+  elementClass=elementClass labelClass=labelClass
   label=label yieldInLabel=yieldInLabel labelWrapperClass=labelWrapperClass
   help=help cid=id validationIcons=validationIcons canShowErrors=canShowErrors
   inputComponent=this}}
