--- addon/templates/components/html-select.hbs
@@ -1,12 +1,0 @@
-<select onchange={{action "change"}} class="form-control {{class}}" id={{id}} required={{required}} size={{size}} autofocus={{autofocus}} disabled={{disabled}} multiple={{multiple}}>
-{{#if prompt}}
-  <option value="" selected={{eq "" selectedValue}}>
-    {{prompt}}
-  </option>
-{{/if}}
-{{#each content key="@index" as |item|}}
-  <option value="{{get item optionValuePath}}" selected={{eq (get item optionValuePath) selectedValue}}>
-    {{get item optionLabelPath}}
-  </option>
-  {{/each}}
-</select>
