--- app/components/erf-html-select.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-rapid-forms/components/html-select';
